Refactored so engine can be used for multiple games. Basically a customizable wrapper of sorts?
