+++ README.md
@@ -2,3 +2,7 @@
 Chess Sandbox (CS, no pun intended lol) is an implementation of chess in Pygame. In addition, it's going to serve as
 a sandbox for integrations with other CS projects and practices ie. ai, networking, mmr matchmaking, etc.
 Seemingly going to end up being a process
+
+
+### Architecture
+CS is designed to be entirely configurable using the config.json file. These customizations include everything from resizeable screen sizes, tile sizes and custom sprites. Essentially you can change anything that you'd think you'd be able to change. Might be kinda gimmicky but eh why not
